Day_47/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
